uxi: version 0.1.0 (#4)
uxi is a package used to build UXI protocol-compliant game engines easily. A Client is the main representation of a game engine, refer to its documentation for more information. The commands which the engine supports being sent to it from a GUI or other is represented as a Command, refer to its documentation for more details.